Refine main and capability structure
