+++ apps/desktop/src/main/index.ts
@@ -16,7 +16,7 @@ import { assertConfigValid, getAppConfig, getBranding } from './config-loader.ts
 import { applySettingsSideEffects, isSetupComplete } from './settings.ts'
 import { publishNotification } from './session-event-dispatcher.ts'
 import { createPromiseChain, createSingleFlight } from './promise-chain.ts'
-import { configureWorkflowService, startWorkflowService, stopWorkflowService } from './workflow-service.ts'
+import { configureWorkflowService, startWorkflowService, stopWorkflowService } from './workflow/workflow-service.ts'
 import { setRuntimeError, setRuntimeReady } from './runtime-status.ts'
 import {
   configureRuntimeInitialization,
