+++ apps/desktop/src/main/ipc-handlers.ts
@@ -21,7 +21,7 @@ import { syncSessionView } from './session-history-loader.ts'
 import { listRuntimeToolsForResolvedContext } from './runtime-tools.ts'
 import { createSandboxWorkspaceDir } from './runtime-paths.ts'
 import { createDestructiveConfirmationManager } from './destructive-actions.ts'
-import { listWorkflows as listWorkflowState } from './workflow-store.ts'
+import { listWorkflows as listWorkflowState } from './workflow/workflow-store.ts'
 import { observePerf } from './perf-metrics.ts'
 import { registerAppHandlers } from './ipc/app-handlers.ts'
 import { registerArtifactHandlers } from './ipc/artifact-handlers.ts'
@@ -45,7 +45,7 @@ import {
 } from './capability-tool-discovery.ts'
 import { resolvePrivateSessionArtifactPath } from './ipc-artifact-access.ts'
 import { createIpcRuntimeContext } from './ipc-runtime-context.ts'
-import { getThreadIndexService } from './thread-index-service.ts'
+import { getThreadIndexService } from './thread-index/thread-index-service.ts'
 import { showNativeConfirmation, type NativeConfirmationOptions } from './native-confirmation.ts'
 import { sdkErrorMessage } from './sdk-error.ts'
