+++ apps/desktop/src/main/ipc/session-handlers.ts
@@ -25,7 +25,7 @@ import { normalizeSessionInfo } from '../opencode-adapter.ts'
 import { shortSessionId } from '../log-sanitizer.ts'
 import { log } from '../logger.ts'
 import { ensureRuntimeContextDirectory } from '../runtime-context.ts'
-import { getThreadIndexService } from '../thread-index-service.ts'
+import { getThreadIndexService } from '../thread-index/thread-index-service.ts'
 import { registerSessionActionHandlers } from './session-action-handlers.ts'
 import { registerSessionCommandHandlers } from './session-command-handlers.ts'
 import { registerSessionFileHandlers } from './session-file-handlers.ts'
