+++ apps/desktop/src/main/thread-index/thread-index-schema.ts
