+++ apps/desktop/src/main/update/update-release-source-generic.ts
@@ -1,5 +1,5 @@
 import type { UpdateReleaseSourceDescriptor } from '@open-cowork/shared'
-import type { UpdateReleaseSourceConfig } from './config-types.ts'
+import type { UpdateReleaseSourceConfig } from '../config-types.ts'
 
 export type GenericHttpReleaseSourceConfig = Extract<UpdateReleaseSourceConfig, { kind: 'generic-http' }>
