+++ apps/desktop/src/main/update/update-release-source-github.ts
@@ -1,5 +1,5 @@
 import type { UpdateReleaseSourceDescriptor } from '@open-cowork/shared'
-import type { UpdateReleaseSourceConfig } from './config-types.ts'
+import type { UpdateReleaseSourceConfig } from '../config-types.ts'
 import { parseGithubRepo } from './update-version.ts'
 import { normalizeUpdateChannel } from './update-release-source-generic.ts'
