+++ docs/configuration.md
@@ -683,6 +683,12 @@ Read-only configured subagents are available from both Build and Plan mode.
 Configured subagents with shell, file-write, or ask-only write permissions are
 available from Build mode only, preserving Plan mode's read-only contract.
 
+Workflow setup has one reserved configured-agent contract:
+`workflow-designer`. When workflows are enabled, downstream builds should keep
+that agent id in `agents`; otherwise the Workflows page cannot open setup
+threads. If a downstream build intentionally renames or removes it, update the
+workflow setup policy in code and config together.
+
 Inference overrides map directly to the SDK's `AgentConfig` fields. Any
 unset field inherits session defaults. Use these to route an agent to a
 different model, tune temperature per agent, or cap runaway tool loops
